Refactor environment variable handling to use centralized env module across multiple files
