+++ app/lib/hooks/useMessageParser.ts
@@ -5,7 +5,7 @@ import { useCallback, useState } from 'react';
 import { NO_EXECUTE_ACTION_ANNOTATION, StreamingMessageParser } from '~/lib/runtime/message-parser';
 import { workbenchStore } from '~/lib/stores/workbench';
 import { createScopedLogger } from '~/utils/logger';
-import { env } from '~/env/server';
+import { env } from '~/env/client';
 
 const logger = createScopedLogger('useMessageParser');
 
@@ -58,7 +58,7 @@ export function useMessageParser() {
   const parseMessages = useCallback((messages: Message[], isLoading: boolean) => {
     let reset = false;
 
-    if (env.NODE_ENV === 'development' && !isLoading) {
+    if (env.NEXT_PUBLIC_NODE_ENV === 'development' && !isLoading) {
       reset = true;
       messageParser.reset();
     }
