+++ app/lib/prisma.ts
@@ -1,5 +1,5 @@
 import { PrismaClient } from '@prisma/client';
-import { env } from '~/env';
+import { env } from '~/env/server';
 
 declare global {
   // eslint-disable-next-line no-var
@@ -8,6 +8,6 @@ declare global {
 
 export const prisma = global.prisma || new PrismaClient();
 
-if (env.server.NODE_ENV !== 'production') {
+if (env.NODE_ENV !== 'production') {
   global.prisma = prisma;
 }
