+++ app/lib/runtime/action-runner.ts
@@ -10,7 +10,7 @@ import { LiblabShell } from '~/utils/shell';
 import { injectEnvVariable } from '~/utils/envUtils';
 import { webcontainer as webcontainerPromise } from '~/lib/webcontainer';
 import { workbenchStore } from '~/lib/stores/workbench';
-import { env } from '~/env';
+import { env } from '~/env/client';
 import { getBaseUrl } from '~/lib/utils/tunnel';
 
 const logger = createScopedLogger('ActionRunner');
@@ -447,11 +447,11 @@ export class ActionRunner {
       let content = action.content;
 
       if (relativePath.endsWith('.env')) {
-        if (env.client.NEXT_PUBLIC_ENV_NAME === 'local') {
+        if (env.NEXT_PUBLIC_ENV_NAME === 'local') {
           const tunnelUrl = await getBaseUrl();
           content = injectEnvVariable(content, 'VITE_API_BASE_URL', tunnelUrl ? tunnelUrl : undefined);
         } else {
-          content = injectEnvVariable(content, 'VITE_API_BASE_URL', env.server.BASE_URL);
+          content = injectEnvVariable(content, 'VITE_API_BASE_URL', env.NEXT_PUBLIC_BASE_URL);
         }
       }
